// src/dtos/generated/user.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class User {
    'id': number;
    'username': string;
    'email': string;
    /**
    * User role (e.g., \'customer\', \'admin\')
    */
    'role': string;
    'created_at'?: string;
    'updated_at'?: string;
    'last_login'?: string;
}